refactor: ran IntelliJ code cleanup

+++ src/main/java/com/mentalfrostbyte/jello/command/impl/HClip.java
@@ -23,7 +23,7 @@ public class HClip extends Command {
                 if (args[0].getCommandType() != CommandType.NUMBER) {
                     throw new CommandException("Invalid distance \"" + args[0].getArguments() + "\"");
                 } else {
-                    float var6 = (float) Math.toRadians((double) (mc.player.rotationYaw + 90.0F));
+                    float var6 = (float) Math.toRadians(mc.player.rotationYaw + 90.0F);
                     double var7 = (double) MathHelper.cos(var6) * args[0].getDouble();
                     double var9 = (double) MathHelper.sin(var6) * args[0].getDouble();
                     mc.getConnection()
