+++ src/main/java/com/mentalfrostbyte/jello/module/impl/movement/highjump/MineplexHighJump.java
@@ -86,14 +86,14 @@ public class MineplexHighJump extends Module {
             double var4 = mc.player.getPosX();
             double var6 = mc.player.getPosZ();
             double var8 = mc.player.getPosY();
-            double var10 = (double) mc.player.movementInput.moveForward;
-            double var12 = (double) mc.player.movementInput.moveForward;
+            double var10 = mc.player.movementInput.moveForward;
+            double var12 = mc.player.movementInput.moveForward;
             float var14 = mc.player.rotationYaw;
             double var15 = 0.1;
             double var17 = var4
-               + (var10 * 0.45 * Math.cos(Math.toRadians((double)(var14 + 90.0F))) + var12 * 0.45 * Math.sin(Math.toRadians((double)(var14 + 90.0F)))) * var15;
+               + (var10 * 0.45 * Math.cos(Math.toRadians(var14 + 90.0F)) + var12 * 0.45 * Math.sin(Math.toRadians(var14 + 90.0F))) * var15;
             double var19 = var6
-               + (var10 * 0.45 * Math.sin(Math.toRadians((double)(var14 + 90.0F))) - var12 * 0.45 * Math.cos(Math.toRadians((double)(var14 + 90.0F)))) * var15;
+               + (var10 * 0.45 * Math.sin(Math.toRadians(var14 + 90.0F)) - var12 * 0.45 * Math.cos(Math.toRadians(var14 + 90.0F))) * var15;
             AxisAlignedBB var21 = new AxisAlignedBB(var17 - 0.3, var8 - 1.0, var19 - 0.3, var17 + 0.3, var8 + 2.0, var19 + 0.3);
             if (mc.world.getCollisionShapes(mc.player, var21).count() == 0L) {
                double var22 = this.method16975(var21);
@@ -106,7 +106,7 @@ public class MineplexHighJump extends Module {
                   mc.getConnection().sendPacket(var26);
                   this.field24026 = var24 + 0.42;
                   mc.player.setPosition(var17, var24, var19);
-                  this.field24025 = (double)this.getNumberValueBySettingName("Motion");
+                  this.field24025 = this.getNumberValueBySettingName("Motion");
                   this.field24024 = 0.81;
                }
             }
